jharkhand-tourism-mvp/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
web3==6.11.0
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.18.0
aiofiles==23.2.1
cors==1.0.1
python-multipart==0.0.6
